+++ dashboard/src/main/home/database-dashboard/DatabaseDashboard.tsx
@@ -19,8 +19,9 @@ import StatusDot from "components/porter/StatusDot";
 import Text from "components/porter/Text";
 import Toggle from "components/porter/Toggle";
 import DashboardHeader from "main/home/cluster-dashboard/DashboardHeader";
+import { isAWSCluster } from "lib/clusters/types";
 import { type ClientDatastore } from "lib/databases/types";
-import { isAWSCluster, useClusterList } from "lib/hooks/useCluster";
+import { useClusterList } from "lib/hooks/useCluster";
 import { useDatastoreList } from "lib/hooks/useDatabaseList";
 
 import { Context } from "shared/Context";
