+++ dashboard/src/main/home/infrastructure-dashboard/tabs/overview/EKSClusterOverview.tsx
@@ -3,23 +3,24 @@ import styled from "styled-components";
 
 import Heading from "components/form-components/Heading";
 import Button from "components/porter/Button";
+import Container from "components/porter/Container";
 import Select from "components/porter/Select";
+import Spacer from "components/porter/Spacer";
 
-type Props = {
-  // Define your component props here
-};
-
-const EKSClusterOverview: React.FC<Props> = (props) => {
+const EKSClusterOverview: React.FC = () => {
   return (
     <>
       <StyledForm>
         <Heading isAtTop>EKS configuration</Heading>
-        <Select
-          options={[]}
-          disabled={true}
-          value={"test"}
-          label="📍 AWS region"
-        />
+        <Spacer y={0.5} />
+        <Container style={{ width: "200px" }}>
+          <Select
+            options={[]}
+            disabled={true}
+            value={"test"}
+            label="📍 AWS region"
+          />
+        </Container>
       </StyledForm>
       <Button
         // disabled={isDisabled()}
